+++ api/src/lidarr/indexer.ts
@@ -1,7 +1,5 @@
 import { Request, Response } from "express";
 
-import { getAppInstance } from "../helpers/app-instance";
-
 import { generateNewznabItem, generateNzbContent } from "./utils/lidarr";
 import { searchTidalForLidarr } from "./utils/tidal-search-albums";
 
@@ -68,7 +66,6 @@ export async function handleSearchRequest(
     return;
   }
 
-  const app = getAppInstance();
   const results = await searchTidalForLidarr(q);
 
   const qualities = ["hires_lossless", "lossless", "high", "low"];